requirements.txt
netmiko
requests
pyyaml
jinja2

